+++ SEMANA7/labefy_proj/src/App.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+
+export default class App extends React.Component{
+  render(){
+    return(
+      <div>
+        <div className= "header"></div>
+        <div className= "footer"></div>
+      </div>
+    )
+  }
+}
